requirements.txt
openai
streamlit
pdfplumber
python-dotenv
